move pica regs
