+++ src/citra_qt/debugger/graphics/graphics_surface.cpp
@@ -18,9 +18,9 @@
 #include "core/hw/gpu.h"
 #include "core/memory.h"
 #include "video_core/pica.h"
-#include "video_core/pica_state.h"
-#include "video_core/regs_framebuffer.h"
-#include "video_core/regs_texturing.h"
+#include "video_core/pica/pica_state.h"
+#include "video_core/pica/regs_framebuffer.h"
+#include "video_core/pica/regs_texturing.h"
 #include "video_core/texture/texture_decode.h"
 #include "video_core/utils.h"
 #include "video_core/video_core.h"
