+++ src/video_core/pica/command_processor.cpp
@@ -16,13 +16,13 @@
 #include "video_core/pica/command_processor.h"
 #include "video_core/debugger/debugger.h"
 #include "video_core/pica.h"
-#include "video_core/pica_state.h"
-#include "video_core/pica_types.h"
+#include "video_core/pica/pica_state.h"
+#include "video_core/pica/pica_types.h"
 #include "video_core/primitive_assembly.h"
 #include "video_core/rasterizer_interface.h"
-#include "video_core/regs.h"
-#include "video_core/regs_pipeline.h"
-#include "video_core/regs_texturing.h"
+#include "video_core/pica/regs.h"
+#include "video_core/pica/regs_pipeline.h"
+#include "video_core/pica/regs_texturing.h"
 #include "video_core/renderer_base.h"
 #include "video_core/shader/shader.h"
 #include "video_core/vertex_loader.h"
