+++ src/video_core/pica/pica_state.h
@@ -10,7 +10,7 @@
 #include "common/vector_math.h"
 #include "video_core/geometry_pipeline.h"
 #include "video_core/primitive_assembly.h"
-#include "video_core/regs.h"
+#include "video_core/pica/regs.h"
 #include "video_core/shader/shader.h"
 
 namespace Pica {
