+++ src/video_core/pica/regs_lighting.h
@@ -5,7 +5,6 @@
 #pragma once
 
 #include <array>
-
 #include "common/assert.h"
 #include "common/bit_field.h"
 #include "common/common_funcs.h"
