+++ src/video_core/pica/regs_shader.h
@@ -5,7 +5,6 @@
 #pragma once
 
 #include <array>
-
 #include "common/bit_field.h"
 #include "common/common_funcs.h"
 #include "common/common_types.h"
