+++ src/video_core/renderer_opengl/gl_rasterizer.cpp
@@ -18,10 +18,10 @@
 #include "core/core.h"
 #include "core/hw/gpu.h"
 #include "video_core/pica.h"
-#include "video_core/pica_state.h"
-#include "video_core/regs_framebuffer.h"
-#include "video_core/regs_rasterizer.h"
-#include "video_core/regs_texturing.h"
+#include "video_core/pica/pica_state.h"
+#include "video_core/pica/regs_framebuffer.h"
+#include "video_core/pica/regs_rasterizer.h"
+#include "video_core/pica/regs_texturing.h"
 #include "video_core/video_core.h"
 #include "video_core/renderer_opengl/gl_rasterizer.h"
 #include "video_core/renderer_opengl/gl_shader_gen.h"
