+++ src/video_core/renderer_opengl/gl_shader_manager.h
@@ -6,7 +6,7 @@
 
 #include <memory>
 #include <glad/glad.h>
-#include "video_core/regs_lighting.h"
+#include "video_core/pica/regs_lighting.h"
 #include "video_core/renderer_opengl/gl_resource_manager.h"
 #include "video_core/renderer_opengl/gl_shader_gen.h"
 #include "video_core/renderer_opengl/gl_state.h"
