+++ src/video_core/swrasterizer/clipper.cpp
@@ -13,8 +13,8 @@
 #include "common/vector_math.h"
 #include "core/core.h"
 #include "video_core/pica.h"
-#include "video_core/pica_state.h"
-#include "video_core/pica_types.h"
+#include "video_core/pica/pica_state.h"
+#include "video_core/pica/pica_types.h"
 #include "video_core/shader/shader.h"
 #include "video_core/video_core.h"
 #include "video_core/swrasterizer/clipper.h"
