+++ src/video_core/swrasterizer/framebuffer.cpp
@@ -12,8 +12,8 @@
 #include "core/hw/gpu.h"
 #include "core/memory.h"
 #include "video_core/pica.h"
-#include "video_core/pica_state.h"
-#include "video_core/regs_framebuffer.h"
+#include "video_core/pica/pica_state.h"
+#include "video_core/pica/regs_framebuffer.h"
 #include "video_core/swrasterizer/framebuffer.h"
 #include "video_core/utils.h"
 #include "video_core/video_core.h"
