+++ src/video_core/swrasterizer/rasterizer.cpp
@@ -18,11 +18,11 @@
 #include "core/hw/gpu.h"
 #include "core/memory.h"
 #include "video_core/pica.h"
-#include "video_core/pica_state.h"
-#include "video_core/pica_types.h"
-#include "video_core/regs_framebuffer.h"
-#include "video_core/regs_rasterizer.h"
-#include "video_core/regs_texturing.h"
+#include "video_core/pica/pica_state.h"
+#include "video_core/pica/pica_types.h"
+#include "video_core/pica/regs_framebuffer.h"
+#include "video_core/pica/regs_rasterizer.h"
+#include "video_core/pica/regs_texturing.h"
 #include "video_core/shader/shader.h"
 #include "video_core/swrasterizer/framebuffer.h"
 #include "video_core/swrasterizer/lighting.h"
