+++ src/video_core/vertex_loader.cpp
@@ -10,9 +10,9 @@
 #include "core/memory.h"
 #include "video_core/debugger/debugger.h"
 #include "video_core/pica.h"
-#include "video_core/pica_state.h"
-#include "video_core/pica_types.h"
-#include "video_core/regs_pipeline.h"
+#include "video_core/pica/pica_state.h"
+#include "video_core/pica/pica_types.h"
+#include "video_core/pica/regs_pipeline.h"
 #include "video_core/shader/shader.h"
 #include "video_core/vertex_loader.h"
 #include "video_core/video_core.h"
