+++ src/video_core/vertex_loader.h
@@ -2,7 +2,7 @@
 
 #include <array>
 #include "common/common_types.h"
-#include "video_core/regs_pipeline.h"
+#include "video_core/pica/regs_pipeline.h"
 
 namespace Pica {
